requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
click==8.1.3
dj-database-url==1.2.0
Django==4.1.7
djangorestframework==3.14.0
drf-spectacular==0.26.1
drf-spectacular-sidecar==2023.3.1
inflection==0.5.1
jsonschema==4.17.3
Markdown==3.4.1
mypy-extensions==1.0.0
packaging==23.0
pathlib==1.0.1
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.1.1
pyrsistent==0.19.3
python-decouple==3.8
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
